s1_orbit: return S1A/S1B when parsing filename (#104)
* s1_orbit: return S1A/S1B when parsing filename

This fixes #103 to get the right orbit file.

Currently we are returning "A" or "B" as a "sensor_id", then searching if the orbit file has that identifier (the single letter, which all orbit files have).

This 1. changes the parser to return the full S1A/S1B, 2. calls it the "mission_id" to match the language in the product spec https://sentinels.copernicus.eu/documents/247904/351187/Copernicus_Sentinels_POD_Service_File_Format_Specification

3. does further simplficication and cleanup of logic to make pylance/flake8 errors go away

* add unit test for diffing the missions in one orbit dir

* use fstring to add EOF extension

* codacy trailing whitespace